Refactoring routes and amend tests.
