+++ frontend/src/app/app-routing.module.ts
@@ -1,14 +1,11 @@
 import { NgModule } from "@angular/core";
 import { Routes, RouterModule } from "@angular/router";
-import { routes as helloRoutes } from "./hello/routes";
 import { AppComponent } from "./app.component";
 
 const routes: Routes = [
   {
-    path: "hello",
-    children: [
-      ...helloRoutes,
-    ]
+    path: "",
+    loadChildren: () => import('./hello/hello.module').then(module => module.HelloModule),
   },
   { path: "*", component: AppComponent }
 ];
