+++ frontend/src/app/app.component.html
@@ -1,9 +1,9 @@
 <ul class="app__links">
   <li>
-    <a href="/">home</a>
+    <a [routerLink]="['']">home</a>
   </li>
   <li>
-    <a href="hello">hello</a>
+    <a [routerLink]="['/hello']">hello</a>
   </li>
 </ul>
 <router-outlet></router-outlet>
