+++ frontend/src/app/app.component.ts
@@ -6,3 +6,4 @@ import { Component } from "@angular/core";
   styleUrls: ['./app.component.scss'],
 })
 export class AppComponent {}
+
