+++ frontend/src/app/hello/hello-routing.module.ts
@@ -3,13 +3,13 @@ import { Routes, RouterModule } from "@angular/router";
 import { HelloComponent } from "./hello.component";
 import { DetailComponent } from "./detail/detail.component";
 
-export const routes: Routes = [
-  { path: "", component: HelloComponent },
-  { path: "detail", component: DetailComponent }
+const routes: Routes = [
+  { path: "hello", component: HelloComponent },
+  { path: "hello/detail", component: DetailComponent }
 ];
 
 @NgModule({
-  imports: [RouterModule.forRoot(routes)],
+  imports: [RouterModule.forChild(routes)],
   exports: [RouterModule]
 })
 export class HelloRoutingModule {}
